Move components to page folders, remove nav
