+++ javascript/webapp/src/components/Header.types.ts
@@ -4,6 +4,5 @@
  */
 
 export interface HeaderProps {
-	onMenuClick: (e: React.MouseEvent) => void
 	onSettingsClick: (e: React.MouseEvent) => void
 }
