+++ javascript/webapp/src/components/Layout.tsx
@@ -11,11 +11,9 @@ import { Header } from './Header.js'
 import {
 	Container,
 	Content,
-	FixedContainer,
 	SlidingContainer,
 	StyledSpinner,
 } from './Layout.styles.js'
-import { NavPanel } from './NavPanel.js'
 import { SettingsPanel } from './SettingsPanel.js'
 
 export const Layout: React.FC<
@@ -23,28 +21,16 @@ export const Layout: React.FC<
 		/*nothing*/
 	}>
 > = memo(function Layout({ children }) {
-	const [isMenuOpen, , dismissMenu, toggleMenu] = useFlyoutPanelState()
 	const [isSettingsOpen, openSettings, dismissSettings] = useFlyoutPanelState()
-	const handleMenuClick = useNonPropagatingClickHandler(toggleMenu)
-
 	return (
 		<>
-			<Container onClick={dismissMenu}>
-				<SlidingContainer className={'layout-container'} isOffset={isMenuOpen}>
-					<Header
-						onMenuClick={handleMenuClick}
-						onSettingsClick={openSettings}
-					/>
-					<Suspense fallback={<StyledSpinner />}>
-						<Content className={'layout-content-container'}>{children}</Content>
-					</Suspense>
-				</SlidingContainer>
+			<Container>
+				<Header onSettingsClick={openSettings} />
+				<Suspense fallback={<StyledSpinner />}>
+					<Content className={'layout-content-container'}>{children}</Content>
+				</Suspense>
 				<Footer />
 			</Container>
-			<FixedContainer isOpen={isMenuOpen} onClick={dismissMenu}>
-				<NavPanel onDismiss={dismissMenu} />
-			</FixedContainer>
-
 			{/* Flyout Panels */}
 			<SettingsPanel isOpen={isSettingsOpen} onDismiss={dismissSettings} />
 		</>
