--- javascript/webapp/src/components/NavPanel.styles.ts
@@ -1,23 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import { Link } from 'react-router-dom'
-import styled from 'styled-components'
-
-export const Heading = styled.h2`
-	margin-bottom: 10px;
-`
-
-export const LinkSection = styled.div`
-	margin-left: 10px;
-	margin-top: 20px;
-`
-
-export const ListItem = styled(Link)`
-	font-size: 14pt;
-	display: block;
-`
-export const Container = styled.div`
-	margin: 8px;
-`
