--- javascript/webapp/src/components/NavPanel.types.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export interface NavPanelProps {
-	onDismiss?: () => void
-}
