+++ javascript/webapp/src/pages/DebugPage/DebugPage.hooks.ts
