+++ javascript/webapp/src/pages/DebugPage/DebugPage.md
