+++ javascript/webapp/src/pages/DebugPage/components/ExamplesDropdown.hooks.ts
@@ -7,16 +7,16 @@ import { Workflow } from '@datashaper/workflow'
 import type { IDropdownOption } from '@fluentui/react'
 import { useCallback, useMemo } from 'react'
 
-import aggregatedLookup from '../../public/specs/aggregated-lookup.json'
-import binning from '../../public/specs/binning.json'
-import categorical from '../../public/specs/categorical.json'
-import dropdown from '../../public/specs/dropdown.json'
-import everyOperation from '../../public/specs/every-operation.json'
-import groupby from '../../public/specs/groupby.json'
-import multistepBinarize from '../../public/specs/multistep-binarize.json'
-import sparkbar from '../../public/specs/sparkbar.json'
-import sparkline from '../../public/specs/sparkline.json'
-import spreadhot from '../../public/specs/spreadhot.json'
+import aggregatedLookup from '../../../../public/specs/aggregated-lookup.json'
+import binning from '../../../../public/specs/binning.json'
+import categorical from '../../../../public/specs/categorical.json'
+import dropdown from '../../../../public/specs/dropdown.json'
+import everyOperation from '../../../../public/specs/every-operation.json'
+import groupby from '../../../../public/specs/groupby.json'
+import multistepBinarize from '../../../../public/specs/multistep-binarize.json'
+import sparkbar from '../../../../public/specs/sparkbar.json'
+import sparkline from '../../../../public/specs/sparkline.json'
+import spreadhot from '../../../../public/specs/spreadhot.json'
 import type { ExamplesDropdownProps } from './ExamplesDropdown.types.js'
 
 const specs: Array<WorkflowObject> = [
