+++ javascript/webapp/src/pages/DebugPage/components/InputTables.styles.ts
