+++ javascript/webapp/src/pages/DebugPage/components/InputTables.tsx
