+++ javascript/webapp/src/pages/DebugPage/components/InputTables.types.ts
