+++ javascript/webapp/src/pages/DebugPage/components/Section.styles.ts
