+++ javascript/webapp/src/pages/DebugPage/components/Section.tsx
