+++ javascript/webapp/src/pages/DebugPage/components/Section.types.ts
