+++ javascript/webapp/src/pages/DebugPage/components/StepOutput.tsx
