+++ javascript/webapp/src/pages/DebugPage/components/Table.hooks.ts
