+++ javascript/webapp/src/pages/DebugPage/components/Table.styles.ts
