+++ javascript/webapp/src/pages/DebugPage/components/Table.tsx
