+++ javascript/webapp/src/pages/DebugPage/components/Table.types.ts
